requirements.txt
rich==13.7.1
pillow==10.3.0
typer==0.12.3
